+++ projetotcc-back/src/main/java/br/edu/utfpr/pb/projetotcc/service/SubEtapaService.java
@@ -3,6 +3,6 @@ package br.edu.utfpr.pb.projetotcc.service;
 import br.edu.utfpr.pb.projetotcc.model.SubEtapa;
 
 public interface SubEtapaService
-			extends CrudService<SubEtapa, Integer>{
+			extends CrudService<SubEtapa, Long>{
 
 }
